worker/pyproject.toml
[project]
name = "pymes-worker"
version = "0.1.0"
description = "PYMES Data Strategy - Worker ETL Service"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "PYMES Team" }]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # Data Processing
    "polars>=1.23.0",
    "pandas>=2.2.0",
    "numpy>=2.1.0",
    "fastexcel>=0.10.0",
    "xlsxwriter>=3.2.0",
    
    # BullMQ for Job Queue
    "bullmq>=2.9.0",
    
    # Database
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.0",
    
    # Storage (S3/MinIO)
    "boto3>=1.36.0",
    
    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    
    # Logging
    "structlog>=25.0.0",
    
    # AI / Gemini
    "google-generativeai>=0.8.0",

    # HTTP Client
    "httpx>=0.28.0",
    
    # Redis
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "types-redis>=4.6.0",
    # Testcontainers for integration tests
    "testcontainers[postgres,redis]>=4.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "PTH",   # flake8-use-pathlib
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "B008",    # do not perform function calls in argument defaults (FastAPI Depends)
    "B904",    # raise without from (sometimes intentional)
    "ARG001",  # unused function argument (FastAPI lifespan app param)
    "ARG002",  # unused method argument (BullMQ token param, required by interface)
    "SIM108",  # ternary operator (readability preference)
    "RUF012",  # mutable class attribute (ClassVar not required for this project)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "bullmq.*",
    "boto3.*",
    "botocore.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
